feat(config): add new outbound configuration
Add a new outbound configuration to the `config/outbound.ts` file. This configuration includes a selector type with the tag `PROXY`. The outbounds for this selector are set to `🚀 AUTO` and `☁️ WARP`, with `🚀 AUTO` as the default outbound. Additionally, a new urltest type with the tag `🚀 AUTO` is added, with the `☁️ WARP` outbound and a test URL.

The changes in the `config/_route.ts` file include the addition of a new route object. This object has the `geosite` property set to an array containing the values `bing` and `openai`, and the `outbound` property set to `🇺🇸 US 美国`.

These changes are made to enhance the outbound configuration and provide support for specific geosites.